+++ packages/react-query/CHANGELOG.md
@@ -38,6 +38,13 @@
 - Updated dependencies [837ed18]
   - @suspensive/react@2.0.0-beta.0
 
+## 1.26.0
+
+### Patch Changes
+
+- Updated dependencies [[`43bd74e`](https://github.com/suspensive/react/commit/43bd74e080fe7487fc43f75688c65af594099ef8)]:
+  - @suspensive/react@1.26.0
+
 ## 1.25.3
 
 ### Patch Changes